+++ src/resources/views/layouts/default/public.blade.php
@@ -46,7 +46,7 @@
 </div>
 
 <!-- Scripts -->
-@yield('scripts')
+@stack('scripts')
 <script src="{{ $appshell->useMix ? mix('/js/appshell.js') : asset('/js/appshell.js') }}"></script>
 
 @include('appshell::layouts.default._footer_includes')